inventory_service/inventory_service/requirements.txt
celery==5.4.0
Django==4.2.16
django-celery-beat==2.7.0
django-celery-results
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pika==1.3.2
psycopg2==2.9.10
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
gunicorn==23.0.0
python-dotenv
requests